Completed LeaseAgreement service layer

+++ src/main/java/com/example/realestatemanagment/Repository/LeaseAgreementRepository.java
@@ -3,5 +3,5 @@ package com.example.realestatemanagment.Repository;
 import com.example.realestatemanagment.Models.LeaseAgreement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaseAgreementRepository extends JpaRepository<LeaseAgreement,Integer> {
+public interface LeaseAgreementRepository extends JpaRepository<LeaseAgreement,Long> {
 }
